Added points check for !transfer command

+++ src/main/java/dev/greyferret/ferretbot/logic/ChatLogic.java
@@ -243,11 +243,15 @@ public class ChatLogic {
 			if (split.length == 4 && StringUtils.isNumeric(split[3])) {
 				logger.info("Points transfer initiated by " + event.getLogin() + ", from " + split[1] + " to " + split[2] + " amount " + split[3]);
 				Long sum = Long.parseLong(split[3]);
-				viewerService.removePoints(split[1], sum);
-				streamElementsAPIProcessor.updatePoints(split[1], (sum * -1));
-				viewerService.addPoints(split[2], sum);
-				streamElementsAPIProcessor.updatePoints(split[2], sum);
-				event.sendMessageWithMention("IQ успешно переведены!");
+				boolean updatedPoints = streamElementsAPIProcessor.updatePoints(split[1], (sum * -1));
+				if (updatedPoints) {
+					event.sendMessageWithMention("Недостаточно IQ у первого зрителя!");
+				} else {
+					viewerService.removePoints(split[1], sum);
+					streamElementsAPIProcessor.updatePoints(split[2], sum);
+					viewerService.addPoints(split[2], sum);
+					event.sendMessageWithMention("IQ успешно переведены!");
+				}
 			}
 		}
